pyproject.toml
[tool.poetry]
name = "releaseworkflows"
version = "0.1.0"
description = "CI/CD"
authors = ["Ben <ben.arteii@proton.me>"]
license = "AGPL-3.0-or-later"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
gitpython = "^3.1.42"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
